+++ bin/store/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the CRDT Store project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.3
+### Author
+Jean
+### Updates
+  - ***SQL Constructors***:
+    - Parsing and formatting of Date and Time fields.
+  - ***Test scenarios***:               
+    - Added test scenarios and tokio test assertions
+
+
 ## 0.2.2
 ### Author
 Jean